+++ CHANGELOG.rst
@@ -2,9 +2,13 @@
 CHANGELOG
 =========
 
-unreleased
-==========
+Unpublished
+===========
 
+* Refactored directory list view for significant performance increses
+* Remove thumbnail generation from the directory list view request response cylce
+* Add Django 4.2 support
+* Add thumbnail view for faster visual management of image libraries
 * Fix File.objects.only() query required for deleting user who own files.
 
 2.2.5 (2023-06-11)